replaces image set
